Template: Rename and move the default card thumbnail
Next commit will allow asset overrides.

https://phabricator.endlessm.com/T31415
