+++ template-ui/src/components/Carousel.vue
@@ -23,7 +23,7 @@
 
 <script>
 import { getThumbnail } from 'kolibri-api';
-import defaultThumbnail from '@/components/default-thumb.svg';
+import defaultThumbnail from '@/assets/default-card-thumbnail.svg';
 
 export default {
   name: 'Carousel',
